[feat] repair fetch data
