+++ src/app/layout.tsx
@@ -5,7 +5,7 @@ import { Suspense } from 'react';
 import '@/styles/global.css';
 import { usePathname } from 'next/navigation';
 import Header from '@/layout/Header';
-import FetchData from '@/provider/faker';
+import FetchData from '@/provider/fetch';
 
 export default function RootLayout({
   children,
